requirements.txt
# Banco de dados
psycopg2-binary==2.9.10
python-dotenv==1.0.1

# Supabase (versões compatíveis modernas)
supabase==2.25.0
supabase-auth==2.25.0
supabase-functions==2.25.0

# Interface gráfica
ttkbootstrap==1.7.3
tkcalendar==1.6.1
pillow==9.5.0

# Web (Flask)
Flask==3.0.3
Werkzeug==3.1.3
Jinja2==3.1.4
click==8.1.7
itsdangerous==2.2.0
MarkupSafe==2.1.5
packaging==24.1

# Deploy
gunicorn==23.0.0